pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-frontmatter>=1.0",
    "ollama>=0.6",
    "watchdog>=6.0",
    "sqlite-vec>=0.1",
    "flask>=3.0",
    "waitress>=3.0",
    "flask-cors>=4.0",
    "pywebview>=5.0",
    "fastmcp>=3.1.1",
    "tenacity>=9.1.4",
    "hnswlib>=0.8.0",
    "cryptography>=42.0",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "detect-secrets>=1.5.0",
    "pre-commit>=3.0",
    "pytest-playwright>=0.5.0",
]

[project.scripts]
sb-init = "engine.init_brain:main"
sb-reindex = "engine.reindex:main"
sb-capture = "engine.capture:main"
sb-search = "engine.search:main"
sb-watch = "engine.watcher:main"
sb-check-links = "engine.links:main_check_links"
sb-install = "scripts.install_native:main"
sb-forget = "engine.forget:main"
sb-read = "engine.read:main"
sb-export = "engine.export:main"
sb-anonymize     = "engine.anonymize:main"
sb-update-memory = "engine.ai:main"
sb-recap         = "engine.intelligence:recap_main"
sb-actions       = "engine.intelligence:actions_main"
sb-digest        = "engine.digest:digest_main"
sb-api           = "engine.api:main"
sb-gui           = "engine.gui:main"
sb-mcp-server    = "engine.mcp_server:main"
sb-health        = "engine.health:main"
sb-merge-duplicates = "engine.merge_cli:merge_duplicates_main"
sb-consolidate  = "engine.consolidate:consolidate_main"
sb-backup        = "engine.backup:backup_main"
sb-restore       = "engine.backup:restore_main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.hatch.build.targets.wheel]
packages = ["engine", "scripts"]
force-include = {"engine/gui/static" = "engine/gui/static"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

